+++ src/checks/duplicated_key.rs
@@ -61,7 +61,10 @@ mod tests {
             (
                 LineEntry {
                     number: 1,
-                    file_path: PathBuf::from(".env"),
+                    file: FileEntry {
+                        path: PathBuf::from(".env"),
+                        file_name: ".env".to_string(),
+                    },
                     raw_string: String::from("FOO=BAR"),
                 },
                 None,
@@ -69,13 +72,19 @@ mod tests {
             (
                 LineEntry {
                     number: 2,
-                    file_path: PathBuf::from(".env"),
+                    file: FileEntry {
+                        path: PathBuf::from(".env"),
+                        file_name: ".env".to_string(),
+                    },
                     raw_string: String::from("FOO=BAR"),
                 },
                 Some(Warning::new(
                     LineEntry {
                         number: 2,
-                        file_path: PathBuf::from(".env"),
+                        file: FileEntry {
+                            path: PathBuf::from(".env"),
+                            file_name: ".env".to_string(),
+                        },
                         raw_string: String::from("FOO=BAR"),
                     },
                     String::from("DuplicatedKey: The FOO key is duplicated"),
@@ -92,7 +101,10 @@ mod tests {
             (
                 LineEntry {
                     number: 1,
-                    file_path: PathBuf::from(".env"),
+                    file: FileEntry {
+                        path: PathBuf::from(".env"),
+                        file_name: ".env".to_string(),
+                    },
                     raw_string: String::from("FOO=BAR"),
                 },
                 None,
@@ -100,7 +112,10 @@ mod tests {
             (
                 LineEntry {
                     number: 2,
-                    file_path: PathBuf::from(".env"),
+                    file: FileEntry {
+                        path: PathBuf::from(".env"),
+                        file_name: ".env".to_string(),
+                    },
                     raw_string: String::from("BAR=FOO"),
                 },
                 None,
@@ -116,7 +131,10 @@ mod tests {
             (
                 LineEntry {
                     number: 1,
-                    file_path: PathBuf::from(".env"),
+                    file: FileEntry {
+                        path: PathBuf::from(".env"),
+                        file_name: ".env".to_string(),
+                    },
                     raw_string: String::from("FOO=BAR"),
                 },
                 None,
@@ -124,13 +142,19 @@ mod tests {
             (
                 LineEntry {
                     number: 2,
-                    file_path: PathBuf::from(".env"),
+                    file: FileEntry {
+                        path: PathBuf::from(".env"),
+                        file_name: ".env".to_string(),
+                    },
                     raw_string: String::from("FOO=BAR"),
                 },
                 Some(Warning::new(
                     LineEntry {
                         number: 2,
-                        file_path: PathBuf::from(".env"),
+                        file: FileEntry {
+                            path: PathBuf::from(".env"),
+                            file_name: ".env".to_string(),
+                        },
                         raw_string: String::from("FOO=BAR"),
                     },
                     String::from("DuplicatedKey: The FOO key is duplicated"),
@@ -139,7 +163,10 @@ mod tests {
             (
                 LineEntry {
                     number: 3,
-                    file_path: PathBuf::from(".env"),
+                    file: FileEntry {
+                        path: PathBuf::from(".env"),
+                        file_name: ".env".to_string(),
+                    },
                     raw_string: String::from("BAR=FOO"),
                 },
                 None,
@@ -147,13 +174,19 @@ mod tests {
             (
                 LineEntry {
                     number: 4,
-                    file_path: PathBuf::from(".env"),
+                    file: FileEntry {
+                        path: PathBuf::from(".env"),
+                        file_name: ".env".to_string(),
+                    },
                     raw_string: String::from("BAR=FOO"),
                 },
                 Some(Warning::new(
                     LineEntry {
                         number: 4,
-                        file_path: PathBuf::from(".env"),
+                        file: FileEntry {
+                            path: PathBuf::from(".env"),
+                            file_name: ".env".to_string(),
+                        },
                         raw_string: String::from("BAR=FOO"),
                     },
                     String::from("DuplicatedKey: The BAR key is duplicated"),
@@ -170,7 +203,10 @@ mod tests {
             (
                 LineEntry {
                     number: 1,
-                    file_path: PathBuf::from(".env"),
+                    file: FileEntry {
+                        path: PathBuf::from(".env"),
+                        file_name: ".env".to_string(),
+                    },
                     raw_string: String::from("FOO=BAR"),
                 },
                 None,
@@ -178,13 +214,19 @@ mod tests {
             (
                 LineEntry {
                     number: 2,
-                    file_path: PathBuf::from(".env"),
+                    file: FileEntry {
+                        path: PathBuf::from(".env"),
+                        file_name: ".env".to_string(),
+                    },
                     raw_string: String::from("FOO=BAR"),
                 },
                 Some(Warning::new(
                     LineEntry {
                         number: 2,
-                        file_path: PathBuf::from(".env"),
+                        file: FileEntry {
+                            path: PathBuf::from(".env"),
+                            file_name: ".env".to_string(),
+                        },
                         raw_string: String::from("FOO=BAR"),
                     },
                     String::from("DuplicatedKey: The FOO key is duplicated"),
@@ -193,7 +235,10 @@ mod tests {
             (
                 LineEntry {
                     number: 3,
-                    file_path: PathBuf::from(".env"),
+                    file: FileEntry {
+                        path: PathBuf::from(".env"),
+                        file_name: ".env".to_string(),
+                    },
                     raw_string: String::from("BAR=FOO"),
                 },
                 None,