+++ src/checks/ending_blank_line.rs
@@ -46,7 +46,10 @@ mod tests {
         let mut checker = EndingBlankLineChecker::default();
         let line = LineEntry {
             number: 1,
-            file_path: PathBuf::from(".env"),
+            file: FileEntry {
+                path: PathBuf::from(".env"),
+                file_name: ".env".to_string(),
+            },
             raw_string: String::from("\n"),
         };
 
@@ -58,7 +61,10 @@ mod tests {
         let mut checker = EndingBlankLineChecker::default();
         let line = LineEntry {
             number: 1,
-            file_path: PathBuf::from(".env"),
+            file: FileEntry {
+                path: PathBuf::from(".env"),
+                file_name: ".env".to_string(),
+            },
             raw_string: String::from("\r\n"),
         };
 
@@ -70,7 +76,10 @@ mod tests {
         let mut checker = EndingBlankLineChecker::default();
         let line = LineEntry {
             number: 1,
-            file_path: PathBuf::from(".env"),
+            file: FileEntry {
+                path: PathBuf::from(".env"),
+                file_name: ".env".to_string(),
+            },
             raw_string: String::from("a"),
         };
         let expected = Some(Warning::new(