+++ src/lib.rs
@@ -39,14 +39,14 @@ fn get_args(current_dir: &OsStr) -> clap::ArgMatches {
                 .short("s")
                 .long("skip")
                 .value_name("CHECK_NAME")
-                .help("Skip checks")
+                .help("Skips checks")
                 .multiple(true)
                 .takes_value(true),
         )
         .arg(
             Arg::with_name("show-checks")
                 .long("show-checks")
-                .help("Show list of available checks"),
+                .help("Shows list of available checks"),
         )
         .get_matches()
 }
@@ -132,7 +132,7 @@ pub fn run() -> Result<Vec<Warning>, Box<dyn Error>> {
 
     let mut warnings: Vec<Warning> = Vec::new();
     for file in new_files {
-        let lines = get_lines(&file)?;
+        let lines = get_lines(file)?;
 
         let result = checks::run(lines, &skip_checks);
         warnings.extend(result);
@@ -141,7 +141,7 @@ pub fn run() -> Result<Vec<Warning>, Box<dyn Error>> {
     Ok(warnings)
 }
 
-fn get_lines(fe: &FileEntry) -> io::Result<Vec<LineEntry>> {
+fn get_lines(fe: FileEntry) -> io::Result<Vec<LineEntry>> {
     let mut f = File::open(&fe.path)?;
 
     let reader = BufReader::new(&f);
@@ -156,7 +156,7 @@ fn get_lines(fe: &FileEntry) -> io::Result<Vec<LineEntry>> {
 
         lines.push(LineEntry {
             number,
-            file_path: fe.path.clone(),
+            file: fe.clone(),
             raw_string,
         })
     }
@@ -172,7 +172,7 @@ fn get_lines(fe: &FileEntry) -> io::Result<Vec<LineEntry>> {
         if last_line.as_str() == "\n" {
             lines.push(LineEntry {
                 number: number + 1,
-                file_path: fe.path.clone(),
+                file: fe,
                 raw_string: last_line,
             });
         }